Fixed the Ajax 302 Redirect Issue
The quote request action method is posted to via an ajax request.
Using RedirectToAction with an ajax request will terminate the http request and return a 302 status code. As a result the 302 will have to be handled in the client side javascript code.

* If this was a normal form post the 302 would be automatically handled by the browser.

To call into another controller from a method that is called by ajax, get the controller using the DependencyResolver, create the ControllerContext for it and set any TempData values on the controller itself.